Move HeroWrapper out of components/common
