+++ src/components/common.js
@@ -3,19 +3,16 @@ import _ from 'lodash/fp'
 import * as qs from 'qs'
 import { Fragment, useState } from 'react'
 import FocusLock from 'react-focus-lock'
-import { b, div, h, h1, img, label, span } from 'react-hyperscript-helpers'
+import { b, div, h, img, label, span } from 'react-hyperscript-helpers'
 import { ButtonPrimary } from 'src/components/common/buttons'
 import Clickable from 'src/components/common/Clickable'
 import { IdContainer } from 'src/components/common/IdContainer'
 import Link from 'src/components/common/Link'
 import { spinnerOverlay } from 'src/components/common/spinners'
-import FooterWrapper from 'src/components/FooterWrapper'
 import { icon } from 'src/components/icons'
 import { TextInput } from 'src/components/input'
 import Modal from 'src/components/Modal'
 import { MiniSortable } from 'src/components/table'
-import TopBar from 'src/components/TopBar'
-import landingPageHero from 'src/images/landing-page-hero.jpg'
 import scienceBackground from 'src/images/science-background.jpg'
 import { Ajax } from 'src/libs/ajax'
 import { getEnabledBrand, isRadX } from 'src/libs/brand-utils'
@@ -142,47 +139,6 @@ export const FocusTrapper = ({ children, onBreakout, ...props }) => {
       }
     }, props)
   }, [children])
-}
-
-export const HeroWrapper = ({ showMenu = true, bigSubhead = false, showDocLink = false, children }) => {
-  const brand = getEnabledBrand()
-
-  return h(FooterWrapper, { alwaysShow: true }, [
-    h(TopBar, { showMenu }),
-    div({
-      role: 'main',
-      style: {
-        flexGrow: 1,
-        color: colors.dark(),
-        padding: '3rem 5rem',
-        backgroundColor: '#fafbfd', // This not-quite-white fallback color was extracted from the background image
-        backgroundImage: `url(${landingPageHero})`,
-        backgroundRepeat: 'no-repeat', backgroundSize: '750px', backgroundPosition: 'right 0 top 0'
-      }
-    }, [
-      // width is set to prevent text from overlapping the background image and decreasing legibility
-      h1({ style: { fontSize: 54, width: 'calc(100% - 460px)' } }, [brand.welcomeHeader]),
-      div({
-        style: {
-          margin: '1rem 0', width: 'calc(100% - 460px)', maxWidth: 700, ...(bigSubhead ?
-            { fontSize: 20, lineHeight: '28px' } :
-            { fontSize: 16, lineHeight: 1.5 })
-        }
-      }, [
-        brand.description,
-        showDocLink ?
-          h(Fragment, [
-            ' ',
-            h(Link, {
-              style: { textDecoration: 'underline' },
-              href: 'https://support.terra.bio/hc/en-us',
-              ...Utils.newTabLinkProps
-            }, ['Learn more about Terra.'])
-          ]) : null
-      ]),
-      children
-    ])
-  ])
 }
 
 export const WarningTitle = ({ children, iconSize = 36 }) => {
