+++ src/pages/LandingPage.js
@@ -1,7 +1,8 @@
 import _ from 'lodash/fp'
 import { useEffect, useState } from 'react'
 import { div, h, h2 } from 'react-hyperscript-helpers'
-import { ButtonOutline, ButtonPrimary, Clickable, HeroWrapper, Link } from 'src/components/common'
+import { ButtonOutline, ButtonPrimary, Clickable, Link } from 'src/components/common'
+import { HeroWrapper } from 'src/components/HeroWrapper'
 import { icon } from 'src/components/icons'
 import hexButton from 'src/images/hex-button.svg'
 import terraHero from 'src/images/terra-hero.png'
